refact: Add useTabGroup hook to handle group handlers and state management
